REF: rename LOO helper function modules and update imports
