+++ src/arviz_stats/loo/compare.py
@@ -11,7 +11,7 @@ from scipy.stats import dirichlet, norm
 
 from arviz_stats.base.stats_utils import get_decimal_places_from_se, round_num
 from arviz_stats.loo import loo
-from arviz_stats.loo.helper_loo import _diff_srs_estimator
+from arviz_stats.loo.loo_helper import _diff_srs_estimator
 from arviz_stats.utils import ELPDData
