+++ src/arviz_stats/loo/loo.py
@@ -5,7 +5,7 @@ import warnings
 from arviz_base import rcParams
 from xarray_einstats.stats import logsumexp
 
-from arviz_stats.loo.helper_loo import (
+from arviz_stats.loo.loo_helper import (
     _check_log_jacobian,
     _compute_loo_results,
     _get_log_likelihood_i,
