+++ src/arviz_stats/loo/loo_approximate_posterior.py
@@ -2,7 +2,7 @@
 
 from arviz_base import rcParams
 
-from arviz_stats.loo.helper_loo import (
+from arviz_stats.loo.loo_helper import (
     _check_log_density,
     _check_log_jacobian,
     _prepare_loo_inputs,
