+++ src/arviz_stats/loo/loo_pit.py
@@ -3,7 +3,7 @@
 import xarray as xr
 from arviz_base import convert_to_datatree, extract
 
-from arviz_stats.loo.helper_loo import _get_r_eff
+from arviz_stats.loo.loo_helper import _get_r_eff
 from arviz_stats.utils import get_log_likelihood_dataset
