+++ src/arviz_stats/loo/loo_score.py
@@ -6,7 +6,7 @@ import numpy as np
 from arviz_base import convert_to_datatree, extract
 
 from arviz_stats.base.stats_utils import round_num
-from arviz_stats.loo.helper_loo import (
+from arviz_stats.loo.loo_helper import (
     _get_r_eff,
     _prepare_loo_inputs,
     _validate_crps_input,
