+++ src/arviz_stats/loo/loo_subsample.py
@@ -5,7 +5,8 @@ import xarray as xr
 from arviz_base import rcParams
 from xarray_einstats.stats import logsumexp
 
-from arviz_stats.loo.helper_loo import (
+from arviz_stats.loo.loo_approximate_posterior import loo_approximate_posterior
+from arviz_stats.loo.loo_helper import (
     _check_log_jacobian,
     _compute_loo_results,
     _get_r_eff,
@@ -17,7 +18,6 @@ from arviz_stats.loo.helper_loo import (
     _select_obs_by_indices,
     _warn_pareto_k,
 )
-from arviz_stats.loo.loo_approximate_posterior import loo_approximate_posterior
 from arviz_stats.utils import ELPDData
